pyproject.toml
[project]
name = "mini-wikipedia-rag"
version = "0.1.0"
description = "Example RAG environment using a few wikipedia articles with Llama Index"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "fastapi>=0.121.3",
    "llama-index-embeddings-azure-openai>=0.4.1",
    "llama-index-llms-azure-openai>=0.4.2",
    "llama-index-core>=0.13.0",
    "pandas>=2.0.0",
    "uvicorn>=0.27.0",
    "pyarrow>=15.0.0",
    "huggingface-hub>=0.20.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "httpx>=0.25.0",
    "jupyter>=1.0.0",
    "nbconvert>=7.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
# Filter out known warnings from dependencies
filterwarnings = [
    "ignore::DeprecationWarning:llama_index.*",
    "ignore:There is no current event loop:DeprecationWarning"
]
